pyproject.toml
[project]
name = "kneecap"
version = "0.0.0"
description = "Open Source RSS Generator and mirror"
readme = "README.md"
classifiers = [
    "Framework :: Django",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "django >=5.1.4",
    "djangorestframework >=3.15.2",
    "feedparser >=6.0.11",
    "python-dateutil >=2.9.0",
    "requests >=2.32.3",
    "django-solo >=2.4.0"
]


[project.optional-dependencies]
dev = [
    "ruff >=0.8.4",
    "coverage >=7.6.10",
    "pre-commit >=4.0.1"
]

[tool.ruff]
line-length = 99
exclude = [
    "**/migrations/*.py",
    ".venv",
    ".git",
    ".backburner"
]
